Use transaction depth to implement savepoints in TX server
This change ensures that when running a nested transaction, the existing
transaction ID is used, and the depth is respected when determining if
a transaction is open/closed.
Ideally the nested transaction should include a separate ID and
open/closed status should be tracked there, but I will leave that to
a separate block of work!